+++ API.Tests/Extensions/QueryableExtensionsTests.cs
@@ -6,6 +6,7 @@ using API.Entities;
 using API.Entities.Enums;
 using API.Entities.Metadata;
 using API.Extensions;
+using API.Extensions.QueryExtensions;
 using API.Helpers.Builders;
 using Xunit;
