+++ API/Data/Repositories/ChapterRepository.cs
@@ -7,6 +7,7 @@ using API.DTOs.Metadata;
 using API.DTOs.Reader;
 using API.Entities;
 using API.Extensions;
+using API.Extensions.QueryExtensions;
 using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Microsoft.EntityFrameworkCore;
