+++ API/Data/Repositories/GenreRepository.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using API.DTOs.Metadata;
 using API.Entities;
 using API.Extensions;
+using API.Extensions.QueryExtensions;
 using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Microsoft.EntityFrameworkCore;
