+++ UI/Web/src/app/_models/preferences/writing-style.ts
@@ -2,6 +2,6 @@
  * Mode the user is reading the book in. Not applicable with ReaderMode.Webtoon
  */
 export enum WritingStyle{
-  Vertical = 0,
-  Horizontal = 1
+  Horizontal = 0,
+  Vertical = 1,
 }
