+++ UI/Web/src/app/_models/reading-list.ts
@@ -30,4 +30,8 @@ export interface ReadingList {
      * If this is empty or null, the cover image isn't set. Do not use this externally. 
      */
      coverImage: string;
+     startYear: number;
+     startMonth: number;
+     endYear: number;
+     endMonth: number;
 }
