+++ UI/Web/src/app/bookmark/_components/bookmarks/bookmarks.component.html
@@ -2,7 +2,7 @@
     <h2 title>
         Bookmarks
     </h2>
-    <h6 subtitle>{{series.length}} Series</h6>
+    <h6 subtitle>{{series.length | number}} Series</h6>
 </app-side-nav-companion-bar>
 <app-bulk-operations [actionCallback]="bulkActionCallback"></app-bulk-operations>
 <app-card-detail-layout
