+++ UI/Web/src/app/cards/bulk-operations/bulk-operations.component.html
@@ -4,7 +4,7 @@
             
             <span class="highlight">
                 <i class="fa fa-check me-1" aria-hidden="true"></i>
-                {{selectionCount}} items selected
+                {{selectionCount | number}} items selected
             </span>
 
             <span>
