+++ UI/Web/src/app/collections/_components/all-collections/all-collections.component.html
@@ -2,7 +2,7 @@
     <h2 title>
         Collections
     </h2>
-    <h6 subtitle>{{collections.length}} Items</h6>
+    <h6 subtitle>{{collections.length | number}} Items</h6>
 </app-side-nav-companion-bar>
 <app-card-detail-layout
     [isLoading]="isLoading"
