+++ UI/Web/src/app/want-to-read/_components/want-to-read/want-to-read.component.html
@@ -5,7 +5,7 @@
                 Want To Read
             </h2>
         </ng-container>
-        <h6 subtitle>{{seriesPagination.totalItems}} Series</h6>
+        <h6 subtitle>{{seriesPagination.totalItems | number}} Series</h6>
     </app-side-nav-companion-bar>
 </div>
